separate username getting function from bridge
