+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/TestData.kt
@@ -34,7 +34,7 @@ internal const val VALID_FILE_URL_PATH =
         .plus("ver-6181004c-a6aa-4eda-b145-a5ff73fc4ad0/a/b/index.html")
 internal const val INVALID_FILE_URL_PATH = "https://78d85043-d04f-486a-8212-bf2601cb63a2/js"
 
-internal const val TEST_USER_NAME_PERMISSION_RESULT = "{\"rakuten.miniapp.user.USER_NAME\":\"DENIED\"}"
+internal const val TEST_CUSTOM_PERMISSION_RESULT = "{\"rakuten.miniapp.user.USER_NAME\":\"DENIED\"}"
 internal const val TEST_USER_NAME = "test_user_name"
 
 internal val TEST_MA = MiniAppInfo(
